// src/redux/types/DeceasedDetails.ts
export interface IDeceasedDetails {
    deceasedSurname: string;
    deceasedName: string;
    maritalStatus: string;
    deceasedDateOfDeath: string;
    deceasedIdType: string;
    deceasedIdNumber: string;
    deceasedPassportNumber: string;
    deceasedDlNumber: string;
    deceasedGender: string;
    deceasedRace: string;
    deceasedNationality: string;
    deceasedDateOfBirth: string;
    deceasedAgeAtDeath: string;
    deceasedOccupation: string;
    deceasedResidentialAddress: string;
    deceasedPostalAddress: string;
  }